pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["weifan <p988744@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
docker = "^5.0.3"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
PyYAML = "^6.0"
click = "^8.1.3"
kubernetes = "^24.2.0"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
email-validator = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
